Conditionally show the featured items section.
The featured item section doesn't apply to me, and might not be
useful to everyone. The empty featured section isn't attractive so
let's remove it entirely if there are not featured posts.

I decided to remove the "regular" section head if there are no
featured items since there is no reason to distinguish them. I could
go either way on if that stays excluded or not.